pbl-transformer-main/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
ultralytics